fix: resolve linting and test errors
