+++ frontend/src/providers/Auth.Context.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+import type { AuthUser } from "../types/authUser";
+
+export interface AuthContextType {
+  user: AuthUser | null;
+  setUser: (user: AuthUser | null) => void;
+  isAuthenticated: boolean;
+}
+
+export const AuthContext = createContext<AuthContextType | null>(null);
